feat(web-ui): Migrated the JavaScript static parts
During migration, the file was split in two, the anonymous jQuery function was replaced by an import, a few functions were created to reduce copy-paste, and some type annotations were added.

In the future, it will probably be a good idea to entirely rewrite these files.

Authors of the original file:
Co-Authored-By: Lucas Henry <lucas.henry@enseirb-matmeca.fr>
Co-Authored-By: Aeinot <antoine@carrincazeaux.fr>