build(idea): Renamed the Python SDK to 'Python CTF'
The previous name was different for each python installation, which led to developpers having to rename it on their machines, and having IDEA files in their `git status` results.

For users of rolling distributions, the version number also changes quite often, which breaks the IntelliJ configuration.

To avoid these constraints, the SDK name is now version-agnostic. The change will require all current developers to recreate their SDK, but it shouldn't be necessary in the future.